consolidate padrino instrumentation
